validation: fix system to prevent overfitting
The original validation system used a random subset of the validation
data which only changed on a rerun, then reported only the final loss.
This led to models with good characteristics on one run, then bad ones
on another.  Update the algorithm to use all validation data and report
the average of several final validation losses.